Isolate exception functions to its own runtime module
